+++ app/src/main/java/com/benoitquenaudon/tvfoot/red/app/data/source/BaseMatchesRepository.kt
@@ -1,0 +1,8 @@
+package com.benoitquenaudon.tvfoot.red.app.data.source
+
+import com.benoitquenaudon.tvfoot.red.app.data.entity.Match
+import io.reactivex.Single
+
+interface BaseMatchesRepository {
+  fun loadPage(pageIndex: Int): Single<List<Match>>
+}
