--- app/src/main/java/com/benoitquenaudon/tvfoot/red/app/injection/scope/ScreenScope.kt
@@ -1,8 +1,0 @@
-package com.benoitquenaudon.tvfoot.red.app.injection.scope
-
-import javax.inject.Scope
-import kotlin.annotation.AnnotationRetention.RUNTIME
-
-@Scope
-@Retention(RUNTIME)
-annotation class ScreenScope
