+++ app/src/main/java/com/benoitquenaudon/tvfoot/red/injection/component/ScreenComponent.kt
@@ -1,0 +1,9 @@
+package com.benoitquenaudon.tvfoot.red.injection.component
+
+import com.benoitquenaudon.tvfoot.red.injection.module.ActivityModule
+import com.benoitquenaudon.tvfoot.red.injection.scope.ScreenScope
+import dagger.Subcomponent
+
+@ScreenScope @Subcomponent interface ScreenComponent {
+  fun plus(activityModule: ActivityModule): ActivityComponent
+}
