+++ app/src/main/java/com/benoitquenaudon/tvfoot/red/injection/scope/ActivityScope.kt
@@ -1,0 +1,8 @@
+package com.benoitquenaudon.tvfoot.red.injection.scope
+
+import javax.inject.Scope
+import kotlin.annotation.AnnotationRetention.RUNTIME
+
+@Scope
+@Retention(RUNTIME)
+annotation class ActivityScope
