+++ app/src/main/java/com/benoitquenaudon/tvfoot/red/util/ActivityExt.kt
@@ -2,9 +2,9 @@ package com.benoitquenaudon.tvfoot.red.util
 
 import android.app.Activity
 import com.benoitquenaudon.tvfoot.red.RedApp
-import com.benoitquenaudon.tvfoot.red.app.injection.component.ActivityComponent
-import com.benoitquenaudon.tvfoot.red.app.injection.component.ScreenComponent
-import com.benoitquenaudon.tvfoot.red.app.injection.module.ActivityModule
+import com.benoitquenaudon.tvfoot.red.injection.component.ActivityComponent
+import com.benoitquenaudon.tvfoot.red.injection.component.ScreenComponent
+import com.benoitquenaudon.tvfoot.red.injection.module.ActivityModule
 
 fun Activity.newScreenComponent(): ScreenComponent {
   return RedApp.getApp(this).appComponent.screenComponent()
